requirements.txt
Flask>=3.1.2
python-docx>=1.2.0
requests>=2.32.5
icalendar>=7.0.0
